+++ test/packages/packagesRequiringDoc/index.ts
@@ -1,0 +1,3 @@
+/** Doc */
+export const doc = true;
+export * from "./sub-package/index";