+++ test/packages/packagesRequiringDoc/sub-package/index.ts
@@ -1,0 +1,3 @@
+/** Doc */
+export const subPackageDoc = true;
+export const subPackageNotDoc = true;